tool/requirements.txt
pyserial==3
